+++ src/stdata.c
@@ -32,12 +32,12 @@ extern const char aWriteError[] = "Write error$";
 int16 enableHighlight = 1;
 
 /* Direction/level lookup tables */
-extern const int dirDeltaX[] = {-1, 1, 1, -1, 0, 1, 0, -1, 0};
-extern const int dirDeltaY[] = {1, 1, -1, -1, 1, 0, -1, 0, 0, -8192, -4096};
+extern const int16 dirDeltaX[] = {-1, 1, 1, -1, 0, 1, 0, -1, 0};
+extern const int16 dirDeltaY[] = {1, 1, -1, -1, 1, 0, -1, 0, 0, -8192, -4096};
 /* Per-level grid dimension (max col/row), read as gridLevelSize[level + 3] for
    level 0..4 -> indices [3..7] = {0x400, 0x100, 0x40, 0x10, 4}, matching the x4
    gridBuf hierarchy (level 3 = 16-wide gridBuf2, level 4 = 4-wide gridBuf1). */
-extern const int gridLevelSize[] = {0, 0x1000, 0x2000, 0x400, 0x100, 0x40, 0x10, 4};
+extern const int16 gridLevelSize[] = {0, 0x1000, 0x2000, 0x400, 0x100, 0x40, 0x10, 4};
 
 /* === Group 6 (0x1632-0x1763): Terrain/grid file strings === */
 int16 gridSignature = 0x3232;
@@ -214,11 +214,11 @@ int16 picRow = 0;
 int16 picReadFromFilePtr = 0;
 
 /* Sprite blit state */
-int armPosition = 0;
-int spriteBlitX = 0;
-int spriteBlitY = 0;
-int spriteBlitW = 0;
-int spriteBlitH = 0;
+int16 armPosition = 0;
+int16 spriteBlitX = 0;
+int16 spriteBlitY = 0;
+int16 spriteBlitW = 0;
+int16 spriteBlitH = 0;
 
 /* Pic decode scalar state */
 int16 picReadBufEndPtr = 0;
@@ -395,7 +395,7 @@ extern const struct UnitTypeRemap unitTypeRemapTable[] = {
 };
 
 /* Difficulty level saved from game state */
-int difficultySaved = 1;
+int16 difficultySaved = 1;
 
 /* Mission table - 56 entries defining mission types per theater/tension */
 extern const struct MissionTableEntry missionTable[56] = {
@@ -493,9 +493,9 @@ struct Pilot hallfameBuf[8];
 struct GameComm far *commData;
 struct Game far *gameData;
 SDL_IOStream *fileHandle;
-int selectedPilotIdx;
-int readItemSize;
-int flightUnitCount;
+int16 selectedPilotIdx;
+int16 readItemSize;
+int16 flightUnitCount;
 uint8 joyReady[4];
 uint8 intRegs[12];
 char todayMissStrBuf[0x1D];
