+++ src/stgen.h
@@ -3,9 +3,9 @@
 /* mission generator (stgen.c) */
 #include "inttype.h"
 
-int mystrlen(const char *str);
+int16 mystrlen(const char *str);
 void missionGenerate(void);
 char *getItemCoordStr(int16);
-void buildTargetLabel(int idx);
+void buildTargetLabel(int16 idx);
 
 #endif /* F15_SE2_STGEN */
